JUnit 5 support Move JUnit 4 classes to subpackage. Existing JUnit 4 specific classes remain, are deprecated and refer to the classes in the subpackage. Simplify the Maven pom file a bit. JUnit 4 and 5 dependencies are now optional, import the one(s) needed.
