+++ src/main/java/com/avanza/gs/test/PartitionedPu.java
@@ -15,11 +15,7 @@
  */
 package com.avanza.gs.test;
 
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Properties;
-
+import com.gigaspaces.security.directory.DefaultCredentialsProvider;
 import org.openspaces.core.GigaSpace;
 import org.openspaces.core.cluster.ClusterInfo;
 import org.openspaces.core.properties.BeanLevelProperties;
@@ -27,7 +23,11 @@ import org.openspaces.pu.container.integrated.IntegratedProcessingUnitContainer;
 import org.openspaces.pu.container.integrated.IntegratedProcessingUnitContainerProvider;
 import org.openspaces.pu.container.support.CompoundProcessingUnitContainer;
 import org.springframework.context.ApplicationContext;
-import com.gigaspaces.security.directory.DefaultCredentialsProvider;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Properties;
 
 /**
  * 
@@ -42,18 +42,18 @@ public final class PartitionedPu implements PuRunner {
 	private final Integer numberOfPrimaries;
 	private final Integer numberOfBackups;
 	private final Properties contextProperties = new Properties();
-	private final Map<String, Properties> beanProperies = new HashMap<>();
+	private final Map<String, Properties> beanProperties = new HashMap<>();
 	private final String lookupGroupName;
 	private final boolean autostart;
 	private final ApplicationContext parentContext;
 	private final boolean useAuthentication;
 
-	public PartitionedPu(PartitionedPuConfigurer configurer) {
+	public PartitionedPu(CommonPartitionedPuConfigurer<?> configurer) {
 		this.puXmlPath = configurer.puXmlPath;
 		this.numberOfBackups = configurer.numberOfBackups;
 		this.numberOfPrimaries = configurer.numberOfPrimaries;
 		this.contextProperties.putAll(configurer.contextProperties);
-		this.beanProperies.putAll(configurer.beanProperies);
+		this.beanProperties.putAll(configurer.beanProperties);
 		this.lookupGroupName = configurer.lookupGroupName;
 		this.autostart = configurer.autostart;
 		this.parentContext = configurer.parentContext;
@@ -115,9 +115,7 @@ public final class PartitionedPu implements PuRunner {
 	private BeanLevelProperties createBeanLevelProperties() {
 		BeanLevelProperties beanLevelProperties = new BeanLevelProperties();
 		beanLevelProperties.setContextProperties(contextProperties);
-		for (Map.Entry<String, Properties> beanProperties : beanProperies.entrySet()) {
-			beanLevelProperties.setBeanProperties(beanProperties.getKey(), beanProperties.getValue());
-		}
+		beanProperties.forEach(beanLevelProperties::setBeanProperties);
 		return beanLevelProperties;
 	}
 	
@@ -138,7 +136,7 @@ public final class PartitionedPu implements PuRunner {
 	
 	@Override
 	public GigaSpace getClusteredGigaSpace() {
-		return GigaSpace.class.cast(getPrimaryInstanceApplicationContext(0).getBean(this.gigaSpaceBeanName)).getClustered();
+		return getPrimaryInstanceApplicationContext(0).getBean(GigaSpace.class, this.gigaSpaceBeanName).getClustered();
 	}
 
 	@Override
