+++ src/main/java/com/avanza/gs/test/junit4/PuConfigurers.java
@@ -13,12 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.avanza.gs.test;
+package com.avanza.gs.test.junit4;
 
-/**
- * @deprecated Use {@link com.avanza.gs.test.junit4.PuConfigurers} for JUnit 4
- */
-@Deprecated
+
 public class PuConfigurers {
 
 	public static PartitionedPuConfigurer partitionedPu(String puXmlPath) {
